+++ src/registry/index.ts
@@ -1,17 +1,17 @@
-/* eslint-disable @typescript-eslint/ban-ts-comment */
-import { BaseControllerMethods, Controller, ControllerMethods, ControllerProps, CRUDBase, SubController } from '../';
+import { BaseControllerMethods, Controller, ControllerMethods, ControllerReturnType, CRUDBase } from '../';
 
-type PartialParameter<T extends (...args: unknown[]) => unknown> = Partial<Parameters<T>[0]> extends infer R
-	? R extends undefined
-		? never
-		: R
-	: never;
-type PartialFunction<T extends (...args: unknown[]) => unknown> = PartialParameter<T> extends infer R
-	? (arg0: R) => ReturnType<T>
-	: never;
-
-type PickByType<T, Value> = {
-	[P in keyof T as T[P] extends Value | undefined ? P : never]: T[P];
+type RegisterFunction = <T, C, U extends CRUDBase, S, M, CONT extends ControllerReturnType<T, C, U, S, M>>(
+	controller: ControllerReturnType<T, C, U, S, M>,
+	db: {
+		[method in keyof Omit<ControllerMethods<unknown, never, never>, '$changeServer' | '$clone'>]?: CONT[method];
+	} & {
+		[sk in keyof S]: {
+			[k in keyof CONT[sk] & keyof BaseControllerMethods<unknown>]?: CONT[sk][k];
+		};
+	}
+) => {
+	register: RegisterFunction;
+	handle: () => Map<string, Controller<unknown, unknown, CRUDBase>>;
 };
 
 /**
@@ -20,35 +20,27 @@ type PickByType<T, Value> = {
  * @returns Object with register and handle methods
  */
 export const controllerRegistry = () => {
-	const routes = new Map<string, Controller<unknown, CRUDBase, CRUDBase>>();
+	const routes = new Map<string, Controller<unknown, unknown, CRUDBase>>();
 
 	/**
 	 * Register controller with database handlers
 	 *
-	 * @param c @see {@link Controller}
+	 * @param controller
 	 * @param db Database handlers for controller and sub controllers
-	 * @returns
+	 *
+	 *  @see {@link Controller}
 	 */
-	const register = <CONT extends Controller<unknown, CRUDBase, CRUDBase>>(
-		controller: CONT,
-		db: {
-			[method in keyof ControllerMethods<unknown, CRUDBase, CRUDBase>]?: PartialFunction<CONT[method]>;
-		} & {
-			[sk in keyof PickByType<
-				Omit<CONT, keyof ControllerMethods<unknown, CRUDBase, CRUDBase> | keyof ControllerProps>,
-				ReturnType<SubController<unknown>>
-			>]: {
-				[k in keyof CONT[sk] & keyof BaseControllerMethods<unknown>]: CONT[sk][k];
-			};
-		}
-	) => {
+	const register: RegisterFunction = (controller, db) => {
 		const { index, create, destroy, read, update, ...rest } = db;
 		// Create copy of controller so we won't delete original keys later
 		const copy = { ...controller };
 
 		Object.keys(rest).forEach(k => {
+			if (typeof rest[k] != 'object') return;
+
 			// Remap properties
-			copy[k] = { ...copy[k], ...rest[k] };
+			const { index, read } = rest[k];
+			copy[k] = { ...copy[k], ...rest[k], index, read };
 
 			// Remap key name
 			delete Object.assign(copy, { [copy[k].$url]: copy[k] })[k];