+++ src/template/createTemplate.ts
@@ -1,9 +1,11 @@
 import { CRUDBase, ControllerMethods, ControllerProps } from '..';
 
-type CreateTemplateData<T, C extends CRUDBase, U extends CRUDBase> = Partial<ControllerMethods<T, C, U>>;
-export type TemplateReturn<T, C extends CRUDBase, U extends CRUDBase> = (
+type CreateTemplateData<T, C, U extends CRUDBase> = Partial<
+	Omit<ControllerMethods<T, C, U>, '$changeServer' | '$clone'>
+>;
+export type TemplateReturn<T, C, U extends CRUDBase> = (
 	c: ControllerProps
-) => ControllerMethods<T, C, U>;
+) => Omit<ControllerMethods<T, C, U>, '$changeServer' | '$clone'>;
 
 /**
  * Creates template for controllers
@@ -11,6 +13,8 @@ export type TemplateReturn<T, C extends CRUDBase, U extends CRUDBase> = (
  * @param data Function with {@link ControllerProps} parameter that should return all controller methods
  * @returns Template for controllers
  */
-export const createTemplate = (data: (c: ControllerProps) => CreateTemplateData<unknown, CRUDBase, CRUDBase>) => {
-	return data as <T, C extends CRUDBase, U extends CRUDBase>(c: ControllerProps) => ControllerMethods<T, C, U>;
+export const createTemplate = (data: (c: ControllerProps) => CreateTemplateData<unknown, never, never>) => {
+	return data as <T, C, U extends CRUDBase, ISARR = true>(
+		c: ControllerProps
+	) => Omit<ControllerMethods<T, C, U, ISARR>, '$changeServer' | '$clone'>;
 };