+++ src/template/index.ts
@@ -1,2 +1,3 @@
 export * from './createTemplate';
 export * from './crudTemplate';
+export * from './ipcTemplate';